- Moved `onBlockPlace` for SpawnerBreak mechanic to `events.blockPlaces.SpawnerPlaceListener`. - Renamed `SpawnerListener` to `SpawnerBreakListener` for clarity. - Updated documentation to include SpawnerBreak mechanic details.
