+++ docs/mechanics/spawnerbreak.md
@@ -1,0 +1,28 @@
+---
+description: BedrockBreak mechanic.
+---
+
+# SpawnerBreak
+
+The SpawnerBreak Mechanic allows players to break spawners and have them drop
+
+
+### Break a Spawner and always get the spawner
+
+```yaml
+# Example
+
+Mechanics:
+  spawnerbreak:
+    probability: 1.0
+```
+
+### Break a Spawner but only have a 1% chance to get it
+
+```yaml
+# Example
+
+Mechanics:
+  spawnerbreak:
+    probability: 0.01
+```
