+++ src/main/java/zone/vao/nexoAddon/events/BlockPlaceListener.java
@@ -3,13 +3,14 @@ package zone.vao.nexoAddon.events;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockPlaceEvent;
-import zone.vao.nexoAddon.events.blockBreaks.SpawnerListener;
+import zone.vao.nexoAddon.events.blockBreaks.SpawnerBreakListener;
+import zone.vao.nexoAddon.events.blockPlaces.SpawnerPlaceListener;
 
 public class BlockPlaceListener implements Listener {
 
     @EventHandler
     public void on(BlockPlaceEvent event) {
-        SpawnerListener.onBlockPlace(event);
+        SpawnerPlaceListener.onBlockPlace(event);
     }
 
 }
